debut refacto main
