--- outfile_to_append.txt
@@ -1,1 +1,0 @@
-BASE
