+++ srcs/exit_routines/exit_shell_routine.c
