+++ srcs/minishell_main_routine/is_empty_input.c
@@ -1,19 +1,25 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   TESTPRINTF.c                                       :+:      :+:    :+:   */
+/*   is_empty_input.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: tchobert <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/01/31 19:32:32 by tchobert          #+#    #+#             */
-/*   Updated: 2025/01/31 19:33:08 by tchobert         ###   ########.fr       */
+/*   Created: 2025/02/03 15:45:28 by tchobert          #+#    #+#             */
+/*   Updated: 2025/02/03 15:45:51 by tchobert         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdio.h>
+#include "minishell.h"
 
-int main(void)
+bool	is_empty_input(t_token_list tokenized_input)
 {
-	printf("Hello\n");
-	return (0);
+	const size_t	list_size = ft_lstsize(tokenized_input);
+
+	if (list_size > 2)
+		return (false);
+	if (((t_token *)tokenized_input->content)->token_type == TOKEN_LIST_START
+		&& list_size == 2)
+		return (true);
+	return (false);
 }
