--- test
@@ -1,17 +1,0 @@
-e2e
-files_for_testing_redirections
-includes
-libft
-Makefile
-minishell
-objs
-outfile_to_append.txt
-readline.supp
-README.md
-srcs
-test
-test_minishell.sh
-testprintf
-TESTPRINTF.c
-tests
-Unity
